parse glob options
